+++ Engine/Stulu/src/Stulu/Scripting/Managed/Bindings/Components/SpriteRenderer.h
@@ -2,7 +2,7 @@
 namespace StuluBindings {
 	class SpriteRendererComponent {
 	public:
-		static inline void getTiling(uint32_t go, struct Vector2* tiling) {
+		static inline void getTiling(uint64_t go, struct Vector2* tiling) {
 			auto gameObject = Stulu::GameObject((entt::entity)go, Stulu::Scene::activeScene());
 			if (gameObject.hasComponent<Stulu::SpriteRendererComponent>()) {
 				auto& spriteRen = gameObject.getComponent<Stulu::SpriteRendererComponent>();
@@ -10,7 +10,7 @@ namespace StuluBindings {
 				tiling->y = spriteRen.tiling.y;
 			}
 		}
-		static inline void setTiling(uint32_t go, struct Vector2* tiling) {
+		static inline void setTiling(uint64_t go, struct Vector2* tiling) {
 			auto gameObject = Stulu::GameObject((entt::entity)go, Stulu::Scene::activeScene());
 			if (gameObject.hasComponent<Stulu::SpriteRendererComponent>()) {
 				auto& spriteRen = gameObject.getComponent<Stulu::SpriteRendererComponent>();
@@ -18,7 +18,7 @@ namespace StuluBindings {
 			}
 		}
 
-		static inline void getColor(uint32_t go, struct Vector4* color) {
+		static inline void getColor(uint64_t go, struct Vector4* color) {
 			auto gameObject = Stulu::GameObject((entt::entity)go, Stulu::Scene::activeScene());
 			if (gameObject.hasComponent<Stulu::SpriteRendererComponent>()) {
 				auto& spriteRen = gameObject.getComponent<Stulu::SpriteRendererComponent>();
@@ -28,7 +28,7 @@ namespace StuluBindings {
 				color->w = spriteRen.color.w;
 			}
 		}
-		static inline void setColor(uint32_t go, struct Vector4* color) {
+		static inline void setColor(uint64_t go, struct Vector4* color) {
 			auto gameObject = Stulu::GameObject((entt::entity)go, Stulu::Scene::activeScene());
 			if (gameObject.hasComponent<Stulu::SpriteRendererComponent>()) {
 				auto& spriteRen = gameObject.getComponent<Stulu::SpriteRendererComponent>();
@@ -36,7 +36,7 @@ namespace StuluBindings {
 			}
 		}
 
-		static inline void setTexture(uint32_t go, uint64_t id) {
+		static inline void setTexture(uint64_t go, uint64_t id) {
 			Stulu::Asset asset;
 			if (Stulu::AssetsManager::saveGetAndType(id, asset, Stulu::AssetType::Texture2D)) {
 				auto gameObject = Stulu::GameObject((entt::entity)go, Stulu::Scene::activeScene());
@@ -46,7 +46,7 @@ namespace StuluBindings {
 				}
 			}
 		}
-		static inline uint64_t getTexture(uint32_t go) {
+		static inline uint64_t getTexture(uint64_t go) {
 			auto gameObject = Stulu::GameObject((entt::entity)go, Stulu::Scene::activeScene());
 			if (gameObject.hasComponent<Stulu::SpriteRendererComponent>()) {
 				auto& spriteRen = gameObject.getComponent<Stulu::SpriteRendererComponent>();
